chore:adiciona as tags para patient e move pastas
